requirements.txt
fastapi==0.111.0
psycopg2==2.9.9
python-dotenv==1.0.1
passlib==1.7.4
